Abstract Command implementation
